Extract reading from perf buffer into a separate trait
